Enable tox support
This patch enable tox support on chatbot project.
Initially, it's skipping several pep8 otherwise it won't pass the CI.
Once the patch get merged, it's possible to start remove the ignored
peps.
This will allow tests to be executed locally, before the patch get
submited, and improve the code quality.
This also add the test-requirements.txt and change the travis yaml file
to use it, instead of the pip install commands.